Reorginize project structure. Added basic folders and files relocated
